wasmtime: move `Memory` to its own module
Purely code motion and fixing up imports.
