+++ crates/wasmtime/src/lib.rs
@@ -241,6 +241,7 @@ mod frame_info;
 mod func;
 mod instance;
 mod linker;
+mod memory;
 mod module;
 mod r#ref;
 mod sig_registry;
@@ -257,6 +258,7 @@ pub use crate::frame_info::{FrameInfo, FrameSymbol};
 pub use crate::func::*;
 pub use crate::instance::Instance;
 pub use crate::linker::*;
+pub use crate::memory::*;
 pub use crate::module::Module;
 pub use crate::r#ref::ExternRef;
 pub use crate::store::*;
